+++ interface_examples/argos_types/lib/auto/DeviceLocation.d.ts
@@ -23,6 +23,14 @@ export type DeviceGeolocation = {
      */
     deviceLocationHeading: number;
     /**
+     * If mission center is configured, this is set to the x coordinate (meters) of the device relative to the mission center. Update of mission center will not trigger location update messages. Cannot be set directly, only through deviceLocationLatitude/Longitude.
+     */
+    deviceLocationX?: number;
+    /**
+     * If mission center is configured, this is set to the y coordinate (meters) of the device relative to the mission center. Update of mission center will not trigger location update messages. Cannot be set directly, only through deviceLocationLatitude/Longitude.
+     */
+    deviceLocationY?: number;
+    /**
      * Height relative to choosen reference in meters
      */
     deviceLocationHeight: number;